% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku-analyze.R
\name{analyze_sudoku}
\alias{analyze_sudoku}
\title{Perform some basic diagnostics on a sudoku puzzle}
\usage{
analyze_sudoku(sudoku_matrix)
}
\arguments{
\item{sudoku_matrix}{a sudoku}
}
\description{
This function returns a puzzle's number of clues, naked singles, hidden signles, whether it is solvable, and whether its solution is unique.
}
\examples{
analyze_sudoku(sudoku)
}
