% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.matrix.sudoku}
\alias{as.matrix.sudoku}
\title{Convert a sudoku to a matrix}
\usage{
\method{as.matrix}{sudoku}(x, ...)
}
\arguments{
\item{x}{an object of class sudoku}

\item{...}{additional arguments.}
}
\description{
Convert a sudoku to a matrix
}
