% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrix.R
\name{as_sudoku.matrix}
\alias{as_sudoku.matrix}
\alias{template_grid_code}
\title{Create a SudokuGrid Object}
\usage{
\method{as_sudoku}{matrix}(object)

template_grid_code()
}
\arguments{
\item{object}{Numeric matrix.
Must be nine rows and nine columns,
filled with either integer from 1 to 9 or \code{NA}.}
}
\value{
A \code{sudoku} object.

\code{template_grid_code()} displays template code that can be used
to initialise a sudoku grid.
}
\description{
Create a SudokuGrid Object
}
\examples{
template_grid_code()

m <- simulate_grid(return_type = "matrix")

sudoku_grid <- as_sudoku(m)
sudoku_grid
}
