% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sudoku_solver.R
\name{Backtraking}
\alias{Backtraking}
\title{Backtraking}
\usage{
Backtraking(M)
}
\arguments{
\item{M}{matrice de Sudoku incomplete}
}
\value{
retourne le backtraking/compteur dans solve_sudoku2
}
\description{
Backtraking
}
\examples{
Backtraking(Grille_incomplete(Sudoku(),30))
}
