% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_helpers.R
\docType{data}
\name{boxes}
\alias{boxes}
\title{Intermediate functions}
\format{An object of class \code{matrix} with 9 rows and 9 columns.}
\usage{
boxes
}
\description{
These functions help the sudoku solver solve puzzles!
}
\keyword{datasets}
