% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cant_bes.R
\name{cant_bes_getter}
\alias{cant_bes_getter}
\title{Intermediat functions}
\usage{
cant_bes_getter(sudoku_df)
}
\arguments{
\item{sudoku_df}{an unsolved sudoku, in dataframe form.}
}
\description{
These functions help the sudoku solver solve puzzles!
}
