% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integrity.R
\name{check_integrity}
\alias{check_integrity}
\title{Check the integrity of a sudoku}
\usage{
check_integrity(sudoku_matrix)
}
\arguments{
\item{sudoku_matrix}{a sudku matrix}
}
\description{
This function allows you to check whether a sudoku solution is good
}
\examples{
solved_puzzle <- solve_sudoku(sudoku)
check_integrity(solved_puzzle)
}
