% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku-checks.R
\name{check_sudoku}
\alias{check_sudoku}
\title{Check that a sudoku is a numeric 9x9 matrix.}
\usage{
check_sudoku(sudoku_matrix)
}
\arguments{
\item{sudoku_matrix}{a sudoku matrix.}
}
\description{
Check that a sudoku is a numeric 9x9 matrix.
}
