% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difficulte.R
\name{difficulte}
\alias{difficulte}
\title{Renvoie une matrice du niveau de difficulté demandé}
\usage{
difficulte(niv)
}
\arguments{
\item{niv}{niveau de difficulté : "Facile"=1, "Moyen"=2, "Difficile"=3, "Expert"=4}
}
\value{
affichage du sudoku
}
\description{
Renvoie une matrice du niveau de difficulté demandé
}
