% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sudoku_Difficulte.R
\name{Difficulte2}
\alias{Difficulte2}
\title{Difficulte Sudoku}
\usage{
Difficulte2(D)
}
\arguments{
\item{D}{un string = c(Low, Medium, High, Expert)}
}
\value{
retourne une matrice grille de Sudoku par difficulte
}
\description{
Difficulte Sudoku
}
\examples{
Difficulte2("Low")
}
