% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_checker.R
\name{element_checker}
\alias{element_checker}
\title{Intermediate functions}
\usage{
element_checker(sudoku_df, cant_bes, dimension = c("row", "col", "box"))
}
\arguments{
\item{sudoku_df}{an unsolved sudoku, in dataframe form.}

\item{cant_bes}{all the cant_bes of the current sudoku}

\item{dimention}{row, col, or box}
}
\description{
These functions help the sudoku solver solve puzzles!
}
