% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existeSurBloc.R
\name{existeSurBloc}
\alias{existeSurBloc}
\title{Détermine si le chiffre est déjà présent dans le bloc}
\usage{
existeSurBloc(X, i, l, c)
}
\value{
True ou False
}
\description{
Détermine si le chiffre est déjà présent dans le bloc
}
