% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existeSurColonne.R
\name{existeSurColonne}
\alias{existeSurColonne}
\title{Détermine s'il existe déjà le numéro sur la colonne}
\usage{
existeSurColonne(X, i, k)
}
\arguments{
\item{X}{la matrice}

\item{i}{le nombre}

\item{k}{la colonne}
}
\value{
True ou false
}
\description{
Détermine s'il existe déjà le numéro sur la colonne
}
