% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existeSurLigne.R
\name{existeSurLigne}
\alias{existeSurLigne}
\title{Détermine s'il existe déjà le numéro sur la ligne}
\usage{
existeSurLigne(X, i, k)
}
\arguments{
\item{X}{la matrice}

\item{i}{le nombre}

\item{k}{la ligne}
}
\value{
True ou false
}
\description{
Détermine s'il existe déjà le numéro sur la ligne
}
