% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrix.R
\name{from_matrix}
\alias{from_matrix}
\alias{template_grid_code}
\title{Create a SudokuGrid Object}
\usage{
from_matrix(matrix)

template_grid_code()
}
\arguments{
\item{matrix}{Numeric matrix.
Must be nine rows and nine columns,
filled with either integer from 1 to 9 or \code{NA}.}
}
\value{
A \linkS4class{SudokuGrid} object.

\code{template_grid_code()} displays template code that can be used
to initialise a sudoku grid.
}
\description{
Create a SudokuGrid Object
}
\examples{
m <- matrix(data = c(
  2,NA,NA, 1, 9,NA,NA,NA, 4,
  NA, 1,NA,NA, 8,NA,NA,NA,NA,
  NA,NA, 5,NA,NA,NA, 3,NA,NA,
  6,NA,NA,NA,NA, 3,NA,NA,NA,
  4, 3,NA,NA,NA,NA,NA, 9, 1,
  NA,NA,NA, 2,NA,NA,NA,NA, 8,
  NA,NA, 7,NA,NA,NA, 5,NA,NA,
  NA,NA,NA,NA, 4,NA,NA, 7,NA,
  1,NA,NA,NA, 7, 6,NA,NA, 3
  ), nrow = 9, ncol = 9, byrow = TRUE)
}
