% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sudoku.R
\name{generate_puzzle}
\alias{generate_puzzle}
\title{Puzzle generator}
\usage{
generate_puzzle(clues = 56, unique = TRUE, seed = NULL)
}
\arguments{
\item{clues}{number of non empty cells in the puzzle (must be <= 81 and >= 0).}

\item{unique}{add the requirement that the puzzle must have a unique solution. defaults to TRUE.}

\item{seed}{integer used to create reproducible randomly generated puzzles.  defaults to NULL.}
}
\description{
This function generates sudoku puzzles with a specified number of clues.
}
\examples{
random_puzzle <- generate_puzzle(clues = 32, unique = TRUE, seed = 56)
print_sudoku(random_puzzle)
}
