% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sudoku.R
\name{generate_sudoku}
\alias{generate_sudoku}
\title{Sudoko generator}
\usage{
generate_sudoku(seed = NULL)
}
\arguments{
\item{seed}{integer used to create reproducible randomly generated puzzles.  defaults to NULL.}
}
\description{
This function generates sudoku puzzles
}
\examples{
print_sudoku(sudoku)
solved_sudoku <- solve_sudoku(sudoku)
print_sudoku(solved_sudoku)
}
