% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SudokuGrid.R
\name{get_tile_index}
\alias{get_tile_index}
\alias{get_tile_indices}
\title{Convert Between Row and Tile Indices}
\usage{
get_tile_index(index)

get_tile_indices(index)
}
\arguments{
\item{index}{For \code{get_tile_index()}, an integer scalar from 1 to 9.
For \code{get_tile_indices}, an integer scalar from 1 to 3.}
}
\value{
For \code{get_tile_index()}, an integer scalar from 1 to 3,
indicating the tile corresponding to that row.
For \code{get_tile_indices}, an integer scalar from 1 to 9,
indicating the row corresponding to that tile.
}
\description{
Convert Between Row and Tile Indices
}
