% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grilles_incompletes.R
\name{Grille_incomplete}
\alias{Grille_incomplete}
\title{Matrice d'une grille de Sudoku}
\usage{
Grille_incomplete(M, n)
}
\arguments{
\item{M}{matrice de Sudoku}

\item{n}{nb de valeur a retirer}
}
\value{
retourne une matrice de Sudoku avec n valeurs cachees aleatoirement
}
\description{
Matrice d'une grille de Sudoku
}
\examples{
Grille_incomplete(Sudoku(),30)
}
