% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resolution_Sudoku.R
\name{index_Grille}
\alias{index_Grille}
\title{Premiere étape de resoltion de l'algorithme de Backtracking}
\usage{
index_Grille(M)
}
\arguments{
\item{M}{matrice incomplete de Sudoku}
}
\value{
retourne une matrice dont les cases de valeur manquante sont remplacees par leur nb de solution possible
}
\description{
Premiere étape de resoltion de l'algorithme de Backtracking
}
\examples{
index_Grille(Grille_incomplete(Sudoku(),30))
}
