% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\name{indice}
\alias{indice}
\title{Donne l'indice d'un nombre dans un vecteur}
\usage{
indice(v, vect)
}
\arguments{
\item{v}{un nombre}

\item{vect}{un vecteur}
}
\value{
0 si le nombre n'est pas dans le vecteur, l'indice s'il y est
}
\description{
Donne l'indice d'un nombre dans un vecteur
}
