% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.sudoku}
\alias{is.sudoku}
\title{Verify that a sudoku is a sudoku.}
\usage{
is.sudoku(x)
}
\arguments{
\item{x}{an object of class sudoku.}

\item{...}{additional arguments.}
}
\description{
Verify that a sudoku is a sudoku.
}
