% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_possible.R
\name{is_possible}
\alias{is_possible}
\title{Is a particular value possible?}
\usage{
is_possible(X, i, j, v, depth = 1L)
}
\value{
\code{FALSE} if the value is impossible, \code{TRUE} if possible but not certain.
}
\description{
Is a particular value possible?
}
