% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku-checks.R
\name{is_unique}
\alias{is_unique}
\title{Check if a puzzle has a unique solution}
\usage{
is_unique(sudoku_matrix, return_all_solutions = FALSE, ...)
}
\arguments{
\item{sudoku_matrix}{a sudoku matrix}

\item{return_all_solutions}{TRUE if you want to return all solutions.  FALSE if you want a TRUE/FALSE check on uniqueness.  Defaults to FALSE.  For puzzles with few clues (fewer than 25), attempting to return all solutions can take a long time.}

\item{...}{additional arguments to pass to solver.}
}
\description{
This function allows you to check whether an unsolved sudoku puzzle has a unique solution.
}
\examples{
random_puzzle <- generate_puzzle(clues = 30, unique = TRUE, seed = 56)
is_unique(random_puzzle)
}
