% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_solver.R
\name{logical_solver}
\alias{logical_solver}
\title{Sudoko logic}
\usage{
logical_solver(sudoku_df, verbose = FALSE)
}
\arguments{
\item{sudoku_df}{an unsolved sudoku, in dataframe form.}

\item{verbose}{set to TRUE if you want to print intermediate steps.  Default is FALSE.}
}
\description{
This function attempt to solve a sudoku logically
}
