% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_vecteur.R
\name{perm3}
\alias{perm3}
\title{Permutation par bloc de 3}
\usage{
perm3(c)
}
\arguments{
\item{c}{un vecteur}
}
\value{
retourne un vecteur permute de 3 valeurs
}
\description{
Permutation par bloc de 3
}
\examples{
perm3(c(1,2,3,4,5))
}
