% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.sudoku}
\alias{plot.sudoku}
\title{Plot a sudoku puzzle.}
\usage{
\method{plot}{sudoku}(x, ...)
}
\arguments{
\item{x}{an object of class sudoku.}

\item{...}{additional arguments.}
}
\description{
Plot a sudoku puzzle.
}
