% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SudokuGrid-methods.R
\name{plot.SudokuGrid}
\alias{plot.SudokuGrid}
\alias{plotValue}
\title{Plot the sudoku grid}
\usage{
\method{plot}{SudokuGrid}(x, ...)

plotValue(x, value, ...)
}
\arguments{
\item{x}{A \linkS4class{SudokuGrid} object.}

\item{...}{Arguments passed to and from other methods.}

\item{value}{Integer scalar from 1 to 9. Display only that value.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the sudoku grid
}
\examples{
sudoku_grid <- simulate_grid()
plot(sudoku_grid)
}
