% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_sudoku}
\alias{plot_sudoku}
\title{Plot a sudoku puzzle.}
\usage{
plot_sudoku(x, ...)
}
\arguments{
\item{x}{an object of class sudoku.}

\item{...}{additional arguments.}
}
\description{
Plot a sudoku puzzle.
}
