% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Sudoku.R
\name{plotSudoku_2}
\alias{plotSudoku_2}
\title{Affichage d'une grille de Sudoku}
\usage{
plotSudoku_2(M)
}
\arguments{
\item{M}{une matrice de Sudoku}
}
\value{
plot la grille du Sudoku
}
\description{
Affichage d'une grille de Sudoku
}
\examples{
plotSudoku_2(diag(1:9))
}
