% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible.R
\name{possible}
\alias{possible}
\title{Informe sur la validité du nombre dans la ligne/colonne/bloc}
\usage{
possible(X, row, col, number)
}
\value{
True pour valide ou False pour non valide
}
\description{
Informe sur la validité du nombre dans la ligne/colonne/bloc
}
