% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku-analyze.R
\name{print.analyzed_sudoku}
\alias{print.analyzed_sudoku}
\title{print method for objects of class analyzed_sudoku}
\usage{
\method{print}{analyzed_sudoku}(x, ...)
}
\arguments{
\item{x}{an object of class analyzed_sudoku}

\item{...}{additional parameters to pass to method}
}
\description{
print method for objects of class analyzed_sudoku
}
