% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-sudoku.R
\name{print.sudoku}
\alias{print.sudoku}
\title{Print method for class sudoku}
\usage{
\method{print}{sudoku}(x, ..., H = NULL, hi = NULL, hj = NULL)
}
\arguments{
\item{x}{A puzzle.}

\item{...}{Passed to other methods}

\item{H, hi, hj}{As in \code{\link{print_sudoku}}.}
}
\description{
Print method for class sudoku
}
