% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-sudoku.R
\name{print_sudoku}
\alias{print_sudoku}
\title{Print a sudoku}
\usage{
print_sudoku(X, H = NULL, hi = NULL, hj = NULL, filename = "")
}
\arguments{
\item{X}{Numeric \eqn{9\times 9} matrix with integer values representing
a sudoku puzzle.}

\item{H}{Logical \eqn{9\times 9} matrix with logical values, \code{TRUE} for
\strong{h}ighlighted entries.}

\item{hi, hj}{Alternative to \code{H} to highlight specific cells}

\item{filename}{Where to print. If \code{filename = ""}, the default,
printing is done to the console. Otherwise a filename.}
}
\description{
Print a sudoku
}
\examples{
eg({
  M2 <- sudoku::generateSudoku(Nblank = 25)
  print_sudoku(M2)
})

eg({
  M3 <- sudoku::solveSudoku(M2)
  print_sudoku(M2, M3 != M2)
})

}
