\name{printSudoku}
\alias{printSudoku}
\title{Print a Sudoku Grid to the Terminal.}
\description{Prints a Sudoku grid (a 9x9 matrix) to the terminal.}
\usage{
  printSudoku(z)
}
\arguments{
  \item{z}{A 9x9 numeric matrix, with '0' representing a blank cell.}
}
\value{None; used for side effect.}
\author{David E. Brahm <\email{brahm@alum.mit.edu}>}
\keyword{array}
