% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SudokuGrid-methods.R
\name{replaceCellValues,SudokuGrid-method}
\alias{replaceCellValues,SudokuGrid-method}
\title{Methods for SudokuGrid}
\usage{
\S4method{replaceCellValues}{SudokuGrid}(
  object,
  i,
  j,
  values,
  given = ifelse(identical(length(values), 1L), TRUE, FALSE)
)
}
\arguments{
\item{object}{A \linkS4class{SudokuGrid} object.}

\item{i}{Row index.}

\item{j}{Column index.}

\item{values}{Integer vectors comprising values from 1 to 9.}

\item{given}{Whether the values were given in the starting grid.
Default to \code{TRUE} if a \code{value} has length 1.}
}
\value{
The updated \linkS4class{SudokuGrid} object.
}
\description{
Methods for SudokuGrid
}
\examples{
sudoku_grid <- simulate_grid()
sudoku_grid

replaceCellValues(sudoku_grid, 1, 2, c(1, 2, 3))
}
