% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{replaceCellValues}
\alias{replaceCellValues}
\title{Generics for SudokuGrid}
\usage{
replaceCellValues(object, i, j, values, ...)
}
\arguments{
\item{object}{A \linkS4class{SudokuGrid} object.}

\item{i}{Row index.}

\item{j}{Column index.}

\item{values}{Integer vectors comprising values from 1 to 9.}

\item{...}{Additional arguments passed to and from other methods.}
}
\value{
\code{replaceCellValues()} returns the updated \linkS4class{SudokuGrid} object.
}
\description{
Generics for SudokuGrid
}
\examples{
showMethods("replaceCellValues")
}
