% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sudoku.R
\name{set_sudoku}
\alias{set_sudoku}
\title{Standardize a sudoku}
\usage{
set_sudoku(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
X but NAs for missing values, (not zeroes).
}
\description{
Standardize a sudoku
}
