% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resolution_Sudoku.R
\name{solutions_case}
\alias{solutions_case}
\title{Valeurs d'une case du Sudoku}
\usage{
solutions_case(M, i, j)
}
\arguments{
\item{M}{matrice incomplete de Sudoku}

\item{i}{entier entre 1 et 9}

\item{j}{entier entre 1 et 9}
}
\value{
retourne le nombre de valeurs possibles dans un case cachee du Sudoku
}
\description{
Valeurs d'une case du Sudoku
}
\examples{
solutions_case(Grille_incomplete(Sudoku(),30),2,2)
}
