% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{solve.sudoku}
\alias{solve.sudoku}
\title{Solve a sudoku.}
\usage{
\method{solve}{sudoku}(a, ...)
}
\arguments{
\item{a}{an object of class sudoku.}

\item{...}{additional arguments.}
}
\description{
Solve a sudoku.
}
