% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_almost_solved}
\alias{solve_almost_solved}
\title{Solve an almost-solved \code{X}}
\usage{
solve_almost_solved(X)
}
\arguments{
\item{X}{An 'almost solved' puzzle, i.e. one in which 7 / 9 columns or rows
are solved.}
}
\description{
Solve an almost-solved \code{X}
}
