% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_backtracking.R
\name{solve_backtracking}
\alias{solve_backtracking}
\title{Solve a sudoku by the backtracking algorithm}
\usage{
solve_backtracking(sudoku_df, empties, verbose)
}
\arguments{
\item{sudoku_df}{a sudku dataframe}

\item{empties}{the empty elements}

\item{verbose}{do you want to see the output at every step? (it's a lot!)}

\item{cant_bes}{a list of what each empty element can't be}
}
\description{
This function recursively solves a sudoku
}
