% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sudoku_solver.R
\name{solve_sudoku1}
\alias{solve_sudoku1}
\title{Solver Sudoku}
\usage{
solve_sudoku1(M)
}
\arguments{
\item{M}{matrice de Sudoku incomplete}
}
\value{
retourne la solution sous forme matricielle
}
\description{
Solver Sudoku
}
\examples{
solve_sudoku1(Grille_incomplete(Sudoku(),30))
}
