% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{strwrap_choices}
\alias{strwrap_choices}
\title{Wrap Choices in Cell}
\usage{
strwrap_choices(x)
}
\arguments{
\item{x}{Integer vector of choices comprised of unique values from 1 to 9.}
}
\value{
A character scalar of choices wrapped over multiple lines if needed,
with a maximum of three choices per line.
}
\description{
Wrap Choices in Cell
}
