% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sudoku.R
\docType{data}
\name{sudoku}
\alias{sudoku}
\alias{sudoku2}
\alias{hard_sudoku}
\alias{evil_sudoku}
\title{Sample sudokus}
\format{Each is a matrix with 9 rows and 9 columns.}
\usage{
sudoku

sudoku2

hard_sudoku

evil_sudoku
}
\description{
These are four sample sudokus for testing and experimentation.  s
}
\details{
`sudoku` is a very easy sudoku. `sudoku2` is a somewhat harder sudoku. `hard_sudoku` is a very hard sudoku.  `evil_sudoku` is an extremenly hard sudoku.
}
\keyword{datasets}
