% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokugh-package.R
\docType{package}
\name{sudokugh-package}
\alias{sudokugh}
\alias{sudokugh-package}
\title{sudokugh: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
