% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SudokuGrid-class}
\alias{SudokuGrid-class}
\title{A sudoku grid}
\value{
An object of class \code{SudokuGrid}.
}
\description{
A sudoku grid
}
\section{Slots}{

\describe{
\item{\code{contents}}{tibble.}
}}

\examples{
new("SudokuGrid")
}
