% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sudokuplyr.R
\docType{data}
\name{ind_list}
\alias{ind_list}
\alias{sudoku}
\alias{sudoku2}
\alias{hard_sudoku}
\alias{evil_sudoku}
\title{Sample sudokus}
\format{Each is a matrix with 9 rows and 9 columns.


`ind_list` is a list of indices of elements associated with each element in a sudoku_df (for use by internal functions).}
\usage{
ind_list

sudoku

sudoku2

hard_sudoku

evil_sudoku
}
\description{
There are four sample sudokus for testing and experimentation.
}
\details{
`sudoku` is a very easy sudoku. `sudoku2` is a somewhat harder sudoku. `hard_sudoku` is a very hard sudoku.  `evil_sudoku` is an extremenly hard sudoku.
}
\keyword{datasets}
