% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoplot.R
\name{sudoplot}
\alias{sudoplot}
\title{Affiche la grille de sudoku}
\usage{
sudoplot(X)
}
\arguments{
\item{X}{une matrice de sudoku}
}
\value{
l'affichage du sudoku
}
\description{
Affiche la grille de sudoku
}
