% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku-analyze.R
\name{summary.analyzed_sudoku}
\alias{summary.analyzed_sudoku}
\title{summary method for objects of class analyzed_sudoku}
\usage{
\method{summary}{analyzed_sudoku}(object, ...)
}
\arguments{
\item{object}{an object of class analyzed_sudoku}

\item{...}{additional parameters to pass to method}
}
\description{
summary method for objects of class analyzed_sudoku
}
