% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.sudoku}
\alias{summary.sudoku}
\title{Summarize a sudoku puzzle}
\usage{
\method{summary}{sudoku}(object, ...)
}
\arguments{
\item{object}{an object of class sudoku.}

\item{...}{additional arguments.}
}
\description{
Summarize a sudoku puzzle
}
