% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resolution_Sudoku.R
\name{valeurs_case}
\alias{valeurs_case}
\title{Valeurs admissibles pour une case cachée}
\usage{
valeurs_case(M, i, j)
}
\arguments{
\item{M}{matrice incomplete de Sudoku}

\item{i}{entier entre 1 et 9}

\item{j}{entier entre 1 et 9}
}
\value{
retourne les valeurs admissibles pour une case cachee
}
\description{
Valeurs admissibles pour une case cachée
}
\examples{
valeurs_case(Grille_incomplete(Sudoku(),30),2,1)
}
