% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhichBloc.R
\name{WhichBloc}
\alias{WhichBloc}
\title{Détermine le bloc dans lequel se trouve le chiffre demandé}
\usage{
WhichBloc(X, l, c)
}
\arguments{
\item{X}{la matrice sudoku}

\item{l}{la ligne}

\item{c}{la colonne}
}
\value{
le bloc en question
}
\description{
Détermine le bloc dans lequel se trouve le chiffre demandé
}
