\name{writeSudoku}
\alias{writeSudoku}
\title{Write a Sudoku Grid to a File}
\description{Writes a Sudoku grid (a matrix) to a file.}
\usage{
  writeSudoku(z, fn)
}
\arguments{
  \item{z}{A Sudoku grid.}
  \item{fn}{A filename.}
}
\value{None; used for its side effect.}
\author{David E. Brahm <\email{brahm@alum.mit.edu}>}
\seealso{\code{\link{readSudoku}}}
\keyword{array}
