% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku_full.R
\docType{data}
\name{X}
\alias{X}
\title{Génère un Sudoku aléatoirement}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9 rows and 9 columns.
}
\usage{
X
}
\value{
Un Sudoku 9x9
}
\description{
Génère un Sudoku aléatoirement
}
\keyword{datasets}
