% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\title{Zoom into a square}
\usage{
zoom(M, I, J, check = TRUE)
}
\arguments{
\item{M}{A sudoku.}

\item{I, J}{Single integers, in \eqn{\{-1, 0, 1\}}. The origin is the centre square.}

\item{check}{Run-time assert.}
}
\value{
The submatrix.
}
\description{
Zoom into a square
}
