% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.sudoku}
\alias{as.sudoku}
\title{Convert a matrix to sudoku}
\usage{
as.sudoku(x, ...)
}
\arguments{
\item{x}{a 9x9 numeric matrix.}

\item{...}{additional arguments.}
}
\description{
Convert a matrix to sudoku
}
