% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difficulte.R
\name{difficulte}
\alias{difficulte}
\title{Niveau de difficulté}
\usage{
difficulte(niv)
}
\arguments{
\item{niv}{niveau de difficulté : "Facile", "Moyen", "Difficile", "Expert"}
}
\value{
un sudoku
}
\description{
Niveau de difficulté
}
