% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrix.R
\name{from_matrix}
\alias{from_matrix}
\alias{template_grid_code}
\title{Create a SudokuGrid Object}
\usage{
from_matrix(x)

template_grid_code()
}
\arguments{
\item{x}{Numeric matrix.
Must be nine rows and nine columns,
filled with either integer from 1 to 9 or \code{NA}.}
}
\value{
A \linkS4class{SudokuGrid} object.

\code{template_grid_code()} displays template code that can be used
to initialise a sudoku grid.
}
\description{
Create a SudokuGrid Object
}
\examples{
template_grid_code()

m <- simulate_grid(return_type = "matrix")

sudoku_grid <- from_matrix(m)
sudoku_grid
}
