% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku-generate.R
\name{generate_sudoku}
\alias{generate_sudoku}
\title{Sudoko generator}
\usage{
generate_sudoku(seed = NULL)
}
\arguments{
\item{seed}{integer used to create reproducible randomly generated puzzles.  defaults to NULL.}
}
\description{
This function generates completed sudoku puzzles
}
\examples{
random_puzzle <- generate_sudoku(seed = 56)
print_sudoku(random_puzzle)
}
