% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_click_indice.R
\name{indice}
\alias{indice}
\title{Indice d'un reel}
\usage{
indice(x)
}
\arguments{
\item{x}{un reel}
}
\value{
retourne l'entier le plus proche de x
}
\description{
Indice d'un reel
}
\examples{
indice(1.6)
}
