% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku_checks.R
\name{is_legal}
\alias{is_legal}
\title{Check if a solution is legal}
\usage{
is_legal(sudoku_df)
}
\arguments{
\item{sudoku_df}{a sudoku matrix}
}
\description{
This function allows you to check whether a sudoku solution is good
}
\examples{
solved_puzzle <- solve_sudoku(sudoku)
is_legal(solved_puzzle)
}
