% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SudokuGrid-methods.R
\name{plot_value}
\alias{plot_value}
\alias{plot_choices}
\alias{plot.sudoku}
\alias{plot_value.sudoku}
\title{Plot the sudoku grid}
\usage{
plot_value(x, value, ...)

plot_choices(x, ...)

\method{plot}{sudoku}(x, ...)

\method{plot_value}{sudoku}(x, value, ...)
}
\arguments{
\item{x}{A \code{sudoku} object.}

\item{value}{Integer scalar from 1 to 9. Display only that value.}

\item{...}{Arguments passed to and from other methods.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the sudoku grid
}
\examples{
sudoku_grid <- simulate_grid()
plot(sudoku_grid)
}
