% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SudokuGrid.R
\name{plot.SudokuGrid}
\alias{plot.SudokuGrid}
\title{Plot the sudoku grid}
\usage{
\method{plot}{SudokuGrid}(x, ...)
}
\arguments{
\item{x}{A \linkS4class{SudokuGrid} object.}

\item{...}{Arguments passed to and from other methods.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the sudoku grid
}
\examples{
sudoku_grid <- simulate_grid()
plot(sudoku_grid)
}
