% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Sudoku.R
\name{plotSudoku_2}
\alias{plotSudoku_2}
\title{Affichage d'une grille de Sudoku}
\usage{
plotSudoku_2(M, Z)
}
\arguments{
\item{M}{une matrice de Sudoku}

\item{Z}{une matrice de Sudoku initiale (voir Shiny)}
}
\value{
plot la grille du Sudoku
}
\description{
Affichage d'une grille de Sudoku
}
\examples{
plotSudoku_2(diag(1:9),matrix(c(0,1,1,rep(0,78)),9,9)+diag(1:9))
}
