% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_sudoku}
\alias{print_sudoku}
\title{Print a sudoku puzzle}
\usage{
print_sudoku(x, ...)
}
\arguments{
\item{x}{an object of class sudoku.}

\item{...}{additional arguments.}
}
\description{
Print a sudoku puzzle
}
