% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_sudoku.R
\name{print_sudoku}
\alias{print_sudoku}
\title{Sudoku printer}
\usage{
print_sudoku(sudoku_matrix)
}
\arguments{
\item{sudoku_matrix}{a 9x9 matrix}
}
\description{
This function prints sudoku puzzles.
}
\examples{
print_sudoku(sudoku)
solved_sudoku <- solve_sudoku(sudoku)
print_sudoku(solved_sudoku)
}
