% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SudokuGrid.R
\name{replaceCellValues,SudokuGrid-method}
\alias{replaceCellValues,SudokuGrid-method}
\title{Methods for SudokuGrid}
\usage{
\S4method{replaceCellValues}{SudokuGrid}(
  object,
  i,
  j,
  values,
  given = ifelse(identical(length(values), 1L), TRUE, FALSE)
)
}
\arguments{
\item{object}{A \linkS4class{SudokuGrid} object.}

\item{i}{Row index.}

\item{j}{Column index.}

\item{values}{Integer vectors comprising values from 1 to 9.}

\item{given}{Whether the values were given in the starting grid.
Default to \code{TRUE} if a \code{value} has length 1.}
}
\value{
The updated \linkS4class{SudokuGrid} object.
}
\description{
Methods for SudokuGrid
}
\examples{
m <- matrix(data = c(
  2,NA,NA, 1, 9,NA,NA,NA, 4,
  NA, 1,NA,NA, 8,NA,NA,NA,NA,
  NA,NA, 5,NA,NA,NA, 3,NA,NA,
  6,NA,NA,NA,NA, 3,NA,NA,NA,
  4, 3,NA,NA,NA,NA,NA, 9, 1,
  NA,NA,NA, 2,NA,NA,NA,NA, 8,
  NA,NA, 7,NA,NA,NA, 5,NA,NA,
  NA,NA,NA,NA, 4,NA,NA, 7,NA,
  1,NA,NA,NA, 7, 6,NA,NA, 3
  ), nrow = 9, ncol = 9, byrow = TRUE)

sudoku_grid <- from_matrix(m)

replaceCellValues(sudoku_grid, 1, 2, c(1, 2, 3))
}
