% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve M}
\usage{
solve(M, depth = 3L)
}
\arguments{
\item{M}{An unsolved puzzle. Zeroes or \code{NA} values represent unsolved
values.}

\item{depth}{A nonnegative integer to limit the depth to which \code{\link{solve}}
will be attempted. If negative, a stack overflow may occur.}
}
\value{
The solved puzzle
}
\description{
Solve M
}
