% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku_solver.R
\name{solve_sudoku}
\alias{solve_sudoku}
\title{Sudoko solver}
\usage{
solve_sudoku(sudoku_matrix, verbose = FALSE)
}
\arguments{
\item{sudoku_matrix}{an unsolved sudoku, in matrix form, with NA values for empty cells}

\item{verbose}{set to TRUE if you want to print intermediate steps.  Default is FALSE.}

\item{attempts}{if the program cannot solve with only logic, it will make educated guesses.  How many attempt should it make with educated guesses before giving up?}
}
\description{
This function allows you to solve a sudoku puzzle
}
\examples{
solved_puzzle <- solve_sudoku(sudoku)
solved_puzzle_2 <- solve_sudoku(sudoku2)
}
