% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku_solver.R
\name{solve_sudoku}
\alias{solve_sudoku}
\title{Sudoko solver}
\usage{
solve_sudoku(sudoku_matrix, verbose = FALSE, shuffle = TRUE)
}
\arguments{
\item{sudoku_matrix}{an unsolved sudoku, in matrix form, with NA values for empty cells}

\item{verbose}{set to TRUE if you want to print intermediate steps.  Default is FALSE. Note that setting to TRUE can increase the time it takes to solve a puzzle by an order of magnitude.}

\item{shuffle}{set to TRUE to induce randomness in backtracking.  Default is TRUE because it tends to be a little faster on average and is useful for generating new puzzles.}
}
\description{
This function solves sudoku puzzles.
}
\examples{
print_sudoku(sudoku)
solved_sudoku <- solve_sudoku(sudoku)
print_sudoku(solved_sudoku)
}
